Move MemoryFirebaseSubscriptionRepository to the API layer
