+++ tmail-backend/jmap/extensions-api/src/main/java/com/linagora/tmail/james/jmap/firebase/MemoryFirebaseSubscriptionRepository.java
@@ -13,14 +13,10 @@ import jakarta.inject.Inject;
 
 import org.apache.james.core.Username;
 import org.apache.james.jmap.api.model.TypeName;
-import org.apache.james.user.api.DeleteUserDataTaskStep;
 import org.reactivestreams.Publisher;
 
 import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.Table;
-import com.google.inject.AbstractModule;
-import com.google.inject.Scopes;
-import com.google.inject.multibindings.Multibinder;
 import com.linagora.tmail.james.jmap.model.DeviceClientIdInvalidException;
 import com.linagora.tmail.james.jmap.model.ExpireTimeInvalidException;
 import com.linagora.tmail.james.jmap.model.FirebaseSubscription;
@@ -36,19 +32,6 @@ import scala.jdk.javaapi.CollectionConverters;
 import scala.jdk.javaapi.OptionConverters;
 
 public class MemoryFirebaseSubscriptionRepository implements FirebaseSubscriptionRepository {
-
-    public static class Module extends AbstractModule {
-        @Override
-        protected void configure() {
-            bind(MemoryFirebaseSubscriptionRepository.class).in(Scopes.SINGLETON);
-            bind(FirebaseSubscriptionRepository.class).to(MemoryFirebaseSubscriptionRepository.class);
-
-            Multibinder.newSetBinder(binder(), DeleteUserDataTaskStep.class)
-                .addBinding()
-                .to(FirebaseSubscriptionUserDeletionTaskStep.class);
-        }
-    }
-
     private final Table<Username, FirebaseSubscriptionId, FirebaseSubscription> table;
     private final Clock clock;
